fix(vtt): combat panel honors percent-vs-bar hp mode; row extracted to meet size cap
